Test cleanup and copyright errors.
